requirements.txt
slackclient==1.3.2
wheel==0.37.1
pandas==1.5.1
python-dotenv==0.21.0